Improved Analytics overview test fixture
no ref

The composed overview scenario now lives above the endpoint helpers and uses the empty responses that exposed the regression, avoiding synthetic dated data.
